ADD CHECKOUT FUNCTIONALITY TO DISK INVENTORY APPLICATION
Building on the last project, add a link to the navigation & home page to allow users to add checkout data. Allow the user to select a disk, a borrower & a checkout date that gets stored in your inventory db. Returned date is optional for both add & edit. Verify the user enters valid dates.